comment_golem_favicon {webR4Shiny}	R Documentation

Comment out favicon

Description

In app_ui.R, golem uses favicon, which does not work well with webR, likely because of the fs package (to check). This function will comment this line so the app UI prints correctly.

Usage

comment_golem_favicon(path)

[Package webR4Shiny version 0.1.0 Index]