set_app_deps {webR4Shiny}	R Documentation

Get app package dependencies to webr-shiny.js

Description

This allows to programmatically write all the necessaries app dependencies to the JS file and load them with library. This is necessary because we can't install the local package from source so we can't benefit from the NAMESPACE (imports).

Usage

set_app_deps(path, conn)

Arguments

path
Path containing the Shiny webR installation.

conn
webr-shiny.js content.

[Package webR4Shiny version 0.1.0 Index]