update_shiny_webr {webR4Shiny}	R Documentation

Update a webR compatible scaffold

Description

Run each time you change any of the R supporting file or inst elements.

Usage

update_shiny_webr(path = "./webr", copy_app_files = TRUE)

Arguments

path
Path to setup the webR installation.

copy_app_files
Whether to copy the Shiny app package main assets (R folder, inst, NAMESPACE and DESCRIPTION). Default to TRUE. If FALSE, it will only copy the webR JS assets and html elements.

[Package webR4Shiny version 0.1.0 Index]